Move tests for `removeAll` into separate file
